+++ ServoCommander/MainWindow.xaml
@@ -22,6 +22,7 @@
                     <RowDefinition Height="60"/>
                     <RowDefinition Height="30"/>
                     <RowDefinition Height="*"/>
+                    <RowDefinition Height="35"/>
                 </Grid.RowDefinitions>
                 <Grid Name="gridConnection" Grid.Row="0"  Background="Aqua">
                     <Label Content="連接埠:" HorizontalAlignment="Left" Margin="10,2,0,0" VerticalAlignment="Top" FontWeight="Bold"/>
@@ -39,36 +40,13 @@
                     <Button x:Name="btnNetConnect" Content="網路連線" HorizontalAlignment="Left" Margin="220,30,0,0" VerticalAlignment="Top" Width="70" Click="btnNetConnect_Click" FontWeight="Bold" Height="23"/>
                 </Grid>
                 <Grid Grid.Row="1">
-                    <RadioButton x:Name="rbControlBoard" GroupName="CmdType" Content="控制板協定" HorizontalAlignment="Left" Margin="10,10,0,0" VerticalAlignment="Top" IsChecked="True" Checked="rbCommand_Checked"/>
-                    <RadioButton x:Name="rbUBTech" GroupName="CmdType" Content="優必選協定" HorizontalAlignment="Left" Margin="110,10,0,0" VerticalAlignment="Top" Checked="rbCommand_Checked"/>
+                    <RadioButton x:Name="rbControlBoard" GroupName="CmdType" Content="控制板協定" HorizontalAlignment="Left" Margin="10,10,0,0" VerticalAlignment="Top" Checked="rbCommand_Checked"/>
+                    <RadioButton x:Name="rbUBTech" GroupName="CmdType" Content="優必選協定" HorizontalAlignment="Left" Margin="110,10,0,0" VerticalAlignment="Top" Checked="rbCommand_Checked" IsChecked="True" />
                     <RadioButton x:Name="rbHaiLzd" GroupName="CmdType" Content="海靈協定" HorizontalAlignment="Left" Margin="210,10,0,0" VerticalAlignment="Top" Checked="rbCommand_Checked"/>
                 </Grid>
                 <Grid Name="gridCommand" Grid.Row="2"  Background="LightGreen">
-                    <Label Content="舵機編號:" HorizontalAlignment="Left" Margin="10,15,0,0" VerticalAlignment="Top" FontWeight="Bold"/>
-                    <TextBox Name ="txtId" HorizontalAlignment="Left" Height="23" Margin="70,17,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="35" PreviewTextInput="tb_PreviewInteger" PreviewKeyDown="tb_PreviewKeyDown_nospace" MaxLength="2"/>
-                    <Label Content="最大編號:" HorizontalAlignment="Left" Margin="120,15,0,0" VerticalAlignment="Top" FontWeight="Bold"/>
-                    <TextBox Name ="txtMaxId" HorizontalAlignment="Left" Height="23" Margin="180,17,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="35" PreviewTextInput="tb_PreviewInteger" PreviewKeyDown="tb_PreviewKeyDown_nospace" MaxLength="2"/>
-                    <Button Name="btnCheckID" Content="檢查 (_Z)" HorizontalAlignment="Left" Margin="222,15,0,0" VerticalAlignment="Top" Width="68" Click="btnCheckID_Click" Height="26" FontWeight="Bold"/>
-                    <RadioButton Name="rbCheckVersion" Content="檢查舵機版本" HorizontalAlignment="Left" Margin="10,50,0,0" VerticalAlignment="Top" IsChecked="True"/>
-                    <RadioButton Name="rbChangeId" Content="更改舵機編號為" HorizontalAlignment="Left" Margin="10,75,0,0" VerticalAlignment="Top" IsChecked="True"/>
-                    <TextBox Name ="txtNewId" HorizontalAlignment="Left" Height="23" Margin="120,70,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="35" PreviewTextInput="tb_PreviewInteger" PreviewKeyDown="tb_PreviewKeyDown_nospace" MaxLength="2"/>
-                    <RadioButton Name="rbGetAngle" Content="讀取舵機角度" HorizontalAlignment="Left" Margin="10,100,0,0" VerticalAlignment="Top"/>
-                    <RadioButton Name="rbMove" Content="移動舵機:" HorizontalAlignment="Left" Margin="10,125,0,0" VerticalAlignment="Top"/>
-                    <Label Content="目標角度: " HorizontalAlignment="Left" Margin="87,120,0,0" VerticalAlignment="Top"  />
-                    <TextBox Name="txtMoveAngle" HorizontalAlignment="Left" Height="23" Margin="145,122,0,0" TextWrapping="NoWrap" Text="" VerticalAlignment="Top" Width="40" PreviewTextInput="tb_PreviewHexMix" PreviewKeyDown="tb_PreviewKeyDown_nospace" CharacterCasing="Upper" MaxLength="4"/>
-                    <Label Content="移動時間: " HorizontalAlignment="Left" Margin="183,120,0,0" VerticalAlignment="Top"/>
-                    <TextBox Name="txtMoveTime" HorizontalAlignment="Left" Height="23" Margin="240,122,0,0" TextWrapping="NoWrap" Text="" VerticalAlignment="Top" Width="50" PreviewTextInput="tb_PreviewHexMix" PreviewKeyDown="tb_PreviewKeyDown_nospace" CharacterCasing="Upper" MaxLength="4"/>
-                    <RadioButton Name="rbGetAdjAngle" Content="讀取舵機角度偏移校正" HorizontalAlignment="Left" Margin="10,150,0,0" VerticalAlignment="Top"/>
-                    <RadioButton Name="rbSetAdjAngle" Content="設定舵機角度偏移量:" HorizontalAlignment="Left" Margin="10,175,0,0" VerticalAlignment="Top" Foreground="#FF780505" FontWeight="Bold"/>
-                    <TextBox Name="txtAdjAngle" HorizontalAlignment="Left" Height="23" Margin="147,172,0,0" TextWrapping="NoWrap" Text="" VerticalAlignment="Top" Width="40" PreviewTextInput="tb_PreviewHex" PreviewKeyDown="tb_PreviewKeyDown_nospace" CharacterCasing="Upper" MaxLength="4" LostFocus="adjAngle_Changed"/>
-                    <TextBox Name="txtAdjMsg" HorizontalAlignment="Left" Height="23" Margin="197,172,0,0" TextWrapping="NoWrap" Text="" VerticalAlignment="Top" Width="93" IsReadOnly="True" Background="LightBlue"/>
-                    <Slider Name="sliderAdjValue" HorizontalAlignment="Left" Margin="29,200,0,0" VerticalAlignment="Top" Width="261" Minimum="-304" Maximum="304" ValueChanged="Slider_ValueChanged" />
-                    <Label Content="成功設定後, 把舵機移動到" HorizontalAlignment="Left" Margin="28,220,0,0" VerticalAlignment="Top"/>
-                    <TextBox Name="txtAdjPreview" HorizontalAlignment="Left" Height="23" Margin="177,222,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="34" PreviewTextInput="tb_PreviewHexMix" PreviewKeyDown="tb_PreviewKeyDown_nospace" CharacterCasing="Upper" MaxLength="4"/>
-                    <Label Content="進行對位." HorizontalAlignment="Left" Margin="218,220,0,0" VerticalAlignment="Top"/>
-                    <RadioButton Name="rbFreeInput" Content="執行以下指令 (不用輸入首尾碼 FA AF 及 ED)" HorizontalAlignment="Left" Margin="10,250,0,0" VerticalAlignment="Top"/>
-                    <TextBox Name="txtCommand" Height="23" Margin="29,270,10,0" TextWrapping="NoWrap" Text="" VerticalAlignment="Top" TextChanged="txtCommand_TextChanged" PreviewTextInput="tb_PreviewCommand" />
-                    <TextBox Name="txtPreview" Height="23" Margin="29,300,10,0" TextWrapping="NoWrap" Text="" VerticalAlignment="Top" IsReadOnly="true" Background="LightBlue"/>
+                </Grid>
+                <Grid Grid.Row="3">
                     <Button Name="btnExecute" Content="發送執行 (_S)" HorizontalAlignment="Left" Margin="5,0,0,5" Width="140" Click="btnExecute_Click" FontWeight="Bold" Height="26" VerticalAlignment="Bottom"/>
                     <Button Name="btnClearLog" Content="清除記錄 (_C)" Margin="0,0,5,5" Click="btnClearLog_Click" FontWeight="Bold" HorizontalAlignment="Right" Width="140" Height="26" VerticalAlignment="Bottom"/>
                 </Grid>
