+++ ServoCommander/uc/UcCommand_ControlBoard.xaml.cs
@@ -24,5 +24,10 @@ namespace ServoCommander.uc
         {
             InitializeComponent();
         }
+
+        public override void ExecuteCommand()
+        {
+            throw new NotImplementedException();
+        }
     }
 }
